requirements.txt
# ── Core PDF parsing ───────────────────────────────────────────────────────
docling>=2.15.0
pymupdf>=1.24.0

# ── Vision & LLM ───────────────────────────────────────────────────────────
google-generativeai>=0.8.0

# ── Vector store ───────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Agent orchestration ────────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

# ── MCP servers ────────────────────────────────────────────────────────────
fastmcp>=2.0.0

# ── API layer ──────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn>=0.30.0

# ── Utilities ──────────────────────────────────────────────────────────────
pydantic>=2.7.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tenacity>=8.3.0
